scraper/requirements.txt
certifi==2020.4.5.1
Scrapy==1.6.0
elasticsearch==7.0.4
python-dotenv==0.13.0
python_dateutil==2.8.1
